+++ src/main.rs
@@ -1,7 +1,10 @@
 use args::parse_args;
 use gw_bin::{
     actions::{script::ScriptAction, Action},
-    checks::{git::GitCheck, Check, CheckError},
+    checks::{
+        git::{CredentialAuth, GitCheck},
+        Check, CheckError,
+    },
     start::{start, StartError},
     triggers::{http::HttpTrigger, once::OnceTrigger, schedule::ScheduleTrigger, Trigger},
 };
@@ -72,7 +75,14 @@ fn main_inner() -> Result<(), MainError> {
 
     // Setup check.
     debug!("Setting up directory {directory} for GitCheck.");
-    let mut check: Box<dyn Check> = Box::new(GitCheck::open(&directory)?);
+    let mut git_check = GitCheck::open(&directory)?;
+    if let Some(ssh_key) = args.ssh_key {
+        git_check.set_auth(CredentialAuth::Ssh(ssh_key));
+    }
+    if let (Some(username), Some(password)) = (args.git_username, args.git_password) {
+        git_check.set_auth(CredentialAuth::Https(username, password));
+    }
+    let mut check: Box<dyn Check> = Box::new(git_check);
 
     // Setup actions.
     let mut actions: Vec<Box<dyn Action>> = vec![];